requirements.txt
aioredis==2.0.1
alembic==1.13.1
aniso8601==10.0.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
astroid==3.0.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
black==23.12.0
blinker==1.9.0
boto3==1.40.62
botocore==1.40.62
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
coverage==7.11.0
cycler==0.12.1
decorator==5.2.1
dill==0.4.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.120.0
flake8==6.1.0
Flask==3.0.0
Flask-Cors==4.0.0
flask-restx==0.5.1
flask-shell-ipython==0.4.1
Flask-Sockets==0.2.1
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.12
psycopg2==2.9.11
pure_eval==0.2.3
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.1.0
Pygments==2.19.2
pylint==3.0.3
pyparsing==3.2.5
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
s3transfer==0.14.0
scikit-learn==1.5.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.19
stack-data==0.6.3
starlette==0.48.0
tabulate==0.9.0
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
python-dateutil